Add to a parent component more to set the information shown in the panel
