+++ lib/frontend/components/base-panelDisplay/index.js
@@ -1,103 +1,50 @@
 import React, { Component } from 'react'
 
+import PanelDisplayData from './panel'
 
-class DisplayDataList extends Component {
-	render () {
-		const item = this.props.items.map( i => {
-			return <nav className="level" key={i.id}>
-				<div className="level-left">
-					<div className="level-item">
-						<p className="subtitle is-5">
-							<strong>{i.text}</strong> 
-						</p>
-					</div>
-				</div>
-				<div className="level-right">
-					<p className="level-item">
-						<button 
-						className="button is-danger"
-						onClick={() => this.props.onRemoveItem(i.id)}>
-							Eliminar
-						</button>
-					</p>
-				</div>
-			</nav>
-		})
+const data = [
+  {id: '4rgsda24', text: 'Text 1'},
+  {id: '5tggeq35', text: 'Text 2'},
+  {id: 'tb52ds51', text: 'Text 3'},
+  {id: 'gg43tsfr', text: 'Text 4'},
+  {id: 'gr3w5dw3', text: 'Text 5'}
+]
 
-		return (
-			<div>{item}</div>
-		)
-	}
+class DisplayData extends Component {
+  constructor () {
+    super()
+    this.state = {
+      value: 'value state',
+      dataList: [],
+     
+    }
+  }   
+
+  AddData () {
+    this.setState({dataList: data})    
+  }
+
+  removeData () {
+    this.setState({dataList: []})
+  }
+
+  render () {
+    return (
+      <div className="columns">
+        <div className="column">
+          <h1>
+            <a className="button " onClick={this.removeData.bind(this)}>Empty</a>
+            <a className="button" onClick={this.AddData.bind(this)}>Data</a>
+            <a className="button ">of Data</a>
+          </h1>
+        </div>
+        <div className="column">
+          <PanelDisplayData data={this.state.dataList}/>
+        </div>
+      </div>
+    )
+  }
+
 }
 
-class PanelDisplayData extends Component {
-	constructor () {
-		super()
-		this.state = {
-			inputValue: '',
-			dataList: []
-		}
-
-		this.handlerChangeData = this.handlerChangeData.bind(this)
-		this.handlerSaveData = this.handlerSaveData.bind(this)
-		this.handlerRemoveData = this.handlerRemoveData.bind(this)
-	}
-
-	handlerChangeData (event) {
-		this.setState({inputValue: event.target.value})
-	}
-
-	handlerSaveData () {
-		var itemData = {
-			id: Math.random().toString(36).substr(2, 16),
-			text: this.state.inputValue
-		}
-
-		this.setState({dataList: this.state.dataList.concat([itemData])})
-	}
-
-	handlerRemoveData (itemId) {
-		const itemRemoveId = this.state.dataList.findIndex(x => x.id === itemId)
-		const removeItem = this.state.dataList.splice(itemRemoveId, 1)
-
-		this.setState({dataList: this.state.dataList})
-	}
-
-	render () {
-		return(
-			<div className="card">
-				<header className="card-header">
-					<p className="card-header-title">
-						Title Panel
-					</p>
-				</header>
-				<div className="card-content">
-					<div className="content">
-						<div className="field is-grouped">
-							<p className="control">
-								<input 
-									className="input" 
-									type="text" 
-									value={this.state.inputValue} 
-									onChange={this.handlerChangeData}
-								/>
-							</p>
-							<p className="control">
-								<button 
-									className="button is-info"
-									onClick={this.handlerSaveData} >
-									Search
-								</button>
-							</p>
-						</div>
-						<DisplayDataList 
-							items={this.state.dataList}
-							onRemoveItem={this.handlerRemoveData}/>
-					</div>
-				</div>
-			</div>
-		)
-	}
-}
-
-export default PanelDisplayData
+export default DisplayData
